Properly split variables/schema/attrs
