+++ catalog_validation/schema/feature_gen.py
@@ -1,0 +1,4 @@
+def get_feature(feature):
+    from .features import FEATURES
+    if feature in FEATURES:
+        return FEATURES[FEATURES.index(feature)]
