+++ catalog_validation/schema/variable_gen.py
@@ -1,0 +1,3 @@
+def generate_variable(variable_data):
+    from .variable import Variable
+    return Variable(variable_data)
