+++ catalog_validation/validation.py
@@ -3,7 +3,7 @@ import yaml
 
 from semantic_version import Version
 
-from catalog_validation.schema.attrs import Variable
+from catalog_validation.schema.variable import Variable
 from .exceptions import CatalogDoesNotExist, ValidationErrors
 from .utils import validate_key_value_types, VALID_TRAIN_REGEX, WANTED_FILES_IN_ITEM_VERSION
